+++ packages/loglevels/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.11] - 2026-01-05
+
+- Added severityNumber property to LogLevelsSet interface for OTLP support.
+- Renamed and refactored LogLevelsDef and LogLevelDef to LogLevelsSet, LogLevelMap and LogLevelSpec.
+
 ## [0.0.10] - 2026-01-05
 
 - Added severityNumber property to LogLevelDef interface for OTLP support